+++ src/cursor-doc/paredit.ts
@@ -266,7 +266,7 @@ export function killForwardList(doc: EditableDocument, start: number = doc.selec
     ], { selection: new ModelEditSelection(start) });
 }
 
-export function forwardSlurpSexp(doc: EditableDocument, start: number = doc.selectionRight) {
+export function forwardSlurpSexp(doc: EditableDocument, start: number = doc.selectionRight, extraOpts = {}) {
     let cursor = doc.getTokenCursor(start);
     cursor.forwardList();
     if (cursor.getToken().type == "close") {
@@ -275,27 +275,45 @@ export function forwardSlurpSexp(doc: EditableDocument, start: number = doc.sele
         cursor.next();
         cursor.forwardSexp();
         cursor.backwardWhitespace(false);
-        doc.model.edit([
-            new ModelEdit('changeRange', [cursor.offsetStart, cursor.offsetStart, close]),
-            new ModelEdit('deleteRange', [offset, 1])
-        ], {});
+        if (cursor.offsetStart !== offset + close.length) {
+            doc.model.edit([
+                new ModelEdit('changeRange', [cursor.offsetStart, cursor.offsetStart, close]),
+                new ModelEdit('deleteRange', [offset, close.length])
+            ], {
+                ...{
+                    undoStopBefore: true
+                },
+                ...extraOpts
+            });
+        } else {
+            forwardSlurpSexp(doc, cursor.offsetStart, { formatParent: true });
+        }
     }
 }
 
-export function backwardSlurpSexp(doc: EditableDocument, start: number = doc.selectionRight) {
+export function backwardSlurpSexp(doc: EditableDocument, start: number = doc.selectionRight, extraOpts = {}) {
     let cursor = doc.getTokenCursor(start);
     cursor.backwardList();
     let tk = cursor.getPrevToken();
     if (tk.type == "open") {
         let offset = cursor.clone().previous().offsetStart;
-        let close = cursor.getPrevToken().raw;
+        let open = cursor.getPrevToken().raw;
         cursor.previous();
         cursor.backwardSexp(true);
         cursor.forwardWhitespace(false);
-        doc.model.edit([
-            new ModelEdit('deleteRange', [offset, tk.raw.length]),
-            new ModelEdit('changeRange', [cursor.offsetStart, cursor.offsetStart, close])
-        ], {});
+        if (offset !== cursor.offsetStart) {
+            doc.model.edit([
+                new ModelEdit('deleteRange', [offset, tk.raw.length]),
+                new ModelEdit('changeRange', [cursor.offsetStart, cursor.offsetStart, open])
+            ], {
+                ...{
+                    undoStopBefore: true
+                },
+                ...extraOpts
+            });
+        } else {
+            backwardSlurpSexp(doc, cursor.offsetStart, { formatParent: true })
+        }
     }
 }
 
@@ -640,7 +658,7 @@ export type WhitespaceInfo = {
 
 /**
  * Collect and return information about the current form regarding its surrounding whitespace
- * @param doc 
+ * @param doc
  * @param p the position in `doc` from where to determine the current form
  */
 export function collectWhitespaceInfo(doc: EditableDocument, p = doc.selectionRight): WhitespaceInfo {